requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# HTTP & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# UI
streamlit>=1.28.0
pandas>=2.0.0

# Utilities
python-multipart>=0.0.6
